Move type_set to a separate file
